requirements.txt
reportlab>=3.6.0
python-barcode>=0.13.1
pillow>=8.0.0